requirements.txt
numpy
pandas
streamlit 
openai
python-dotenv
streamlit-folium
streamlit-geolocation
fastapi==0.115.0
langchain
langchain-openai
langchain-community==0.0.19
unstructured
faiss-cpu